app/pyproject.toml
[project]
name = "my-fast-html-app"
version = "0.1.0"
description = "A FastHTML app"
requires-python = ">=3.12"
dependencies = [
    "python-fasthtml~=0.12.0",
    "monsterui~=1.0.0",
]